# dist/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "FundooHR Backend API's Specification",
        "description": "FundooHR is not a Software product, it is a thinking which make your task easy and efficient. And we are plotting our thining in software product. Which will make our work easy. Basically we are planning to make product based on some different technologies.",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Bridgelabz.com",
            "email": "admin@bridgelabz.com"
        }
    },
    "basePath": "/",
...
    "schemes": [
        "http",
        "https"
    ],
...
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json",
        "application/xml"
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "description": "It authenticates the user email from company database and it returns success response with email and token. For every REST call you need to pass token and crossponding email with some extra request body which is based on REST call's.",
                "operationId": "login",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                        "name": "emailID",
                        "emailID": "Logging EmailID",
                        "in": "formData",
                        "description": "Provide a valid email address to authenticate. e.g. your@some-domain.com",
                        "required": true,
                        "type": "string",
                        "default": "admin@bridgelabz.com"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "Provide a valid Password",
                        "required": true,
                        "type": "string",
                        "format": "password",
                        "default": "Bridge@123"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully authentcation the user",
                        "schema": {
                            "$ref": "#/definitions/loginResponseSuccess"
                        }
                    },
                    "401": {
                        "description": "Failed to authentcation user",
                        "schema": {
                            "$ref": "#/definitions/loginResponseFailure"
                        }
                    }
                }
            }
        },
        "/readDashboardData": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "description": "It return you Dashboard Data",
                "operationId": "dashboardByDate",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                        "name": "date",
                        "date": "Date parameter",
                        "in": "query",
                        "description": "Date parameter to get the Day, month & Year to filter the Dashboard based on month year",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "x-token",
                        "in": "header",
                        "description": "Token needs to validate & authenticate",
                        "required": true,
                        "type": "String"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfull in retriving the Dashboard Data",
                        "schema": {
                            "$ref": "#/definitions/DashboardResponseSuccess"
                        }
                    },
                    "401": {
                        "description": "Failed to retrive the Dashboard Data",
                        "schema": {
                            "$ref": "#/definitions/DashboardResponseFailure"
                        }
                    }
                }
            }
        },
...
        "/readMonthlyAttendanceSummary": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "description": "Summary of attandance for the selected range.",
                "operationId": "attendanceSummaryByDateRange",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "fromDate",
                        "fromDate": "From Date",
                        "in": "query",
                        "description": "From Date parameter to get the Month & Year to get the attendance, Format needs to be in dd/mm/yyyy for .e.g. 01/01/2017",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "toDate",
                        "toDate": "To Date",
                        "in": "query",
                        "description": "To Date parameter to get the Month & Year to get the attendance, Format needs to be in dd/mm/yyyy for .e.g. 01/01/2017",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "x-token",
                        "in": "header",
                        "description": "Token needs to validate & authenticate",
                        "required": true,
                        "type": "String"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success in returning the attendance summary for the selected date range",
                        "schema": {
                            "$ref": "#/definitions/attendanceSummaryByDateRangeSuccess"
                        }
                    },
                    "401": {
                        "description": "Failed to retrive the attendace summary for the selected date range",
                        "schema": {
                            "$ref": "#/definitions/attendanceSummaryByDateRangeFailure"
                        }
                    }
                }
            }
        },
...
        "/readEmployeeMonthlyAttendance": {
            "get": {
                "tags": [
                    "Attendance", "Engineer"
                ],
                "description": "Recorded attendance for the employee in a month",
                "operationId": "employeeMonthAttendance",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "engineerID",
                        "in": "query",
                        "description": "Engineer ID is unique for each engineer",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "date",
                        "date": "Date to extract month",
                        "in": "query",
                        "description": "Date parameter to get the Month & Year. Format needs to be in dd/mm/yyyy for .e.g. 01/01/2017",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "x-token",
                        "in": "header",
                        "description": "Token needs to validate & authenticate",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success in getting employee attendance record for a month",
                        "schema": {
                            "$ref": "#/definitions/employeeMonthAttendanceSuccess"
                        }
                    },
                    "401": {
                        "description": "Failed to get the employee attendance record for a month",
                        "schema": {
                            "$ref": "#/definitions/employeeMonthAttendanceFailure"
                        }
                    }
                }
            }
        },
        "/createEmployeeDayAttendance": {
            "post": {
                "tags": [
                    "Attendance",
                    "Engineer"
                ],
                "description": "Create the employee attendance record",
                "operationId": "recordEmployeeDayAttendance",
                "produces": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "Scehma",
                        "in": "body",
                        "description": "JSON format for sending the employee details & attendance details.",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/recordEmployeeDayAttendanceSchema"
                        }
                    },
                    {
                        "name": "x-token",
                        "in": "header",
                        "description": "Token needs to validate & authenticate",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success in getting employee attendance record for a month",
                        "schema": {
                            "$ref": "#/definitions/recordEmployeeDayAttendanceSuccess"
                        }
                    },
                    "401": {
                        "description": "Failed to get the employee attendance record for a month",
                        "schema": {
                            "$ref": "#/definitions/recordEmployeeDayAttendanceFailure"
                        }
                    }
                }
            }
        },
...
        "/readAllEmployee": {
            "get": {
                "tags": [
                    "Engineer",
                    "Report"
                ],
                "description": "Generate the Salary report/invoice of all the employee",
                "operationId": "generateAllEmployeeSalaryReportInvoice",
                "produces": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [{
                    "name": "x-token",
                    "in": "header",
                    "description": "Token needs to validate & authenticate",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "Success in updating Employee Profile Details",
                        "schema": {
                            "$ref": "#/definitions/generateAllEmployeeSalaryReportInvoiceSuccess"
                        }
                    },
                    "401": {
                        "description": "Failed to update Employee Profile Details",
                        "schema": {
                            "$ref": "#/definitions/generateAllEmployeeSalaryReportInvoiceFailure"
                        }
                    }
                }
            }
        },
        "/readInternEmployee": {
            "get": {
                "tags": [
                    "Engineer",
                    "Report"
                ],
                "description": "Internship Employee list to generate attendance reports who are working with tech company",
                "operationId": "attendanceOfInternshipAtClient",
                "produces": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [{
                    "name": "x-token",
                    "in": "header",
                    "description": "Token needs to validate & authenticate",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "Success in retriving attendance of Internship Employee at client side",
                        "schema": {
                            "$ref": "#/definitions/attendanceOfInternshipAtClientSuccess"
                        }
                    },
                    "401": {
                        "description": "Failed to retrive attendance of Internship Employee at client side",
                        "schema": {
                            "$ref": "#/definitions/attendanceOfInternshipAtClientFailure"
                        }
                    }
                }
            }
        },
...
        "/searchEmployeeByName": {
            "get": {
                "tags": [
                    "Engineer", "Search"
                ],
                "description": "Search result for employee by name",
                "operationId": "searchEmployeeListByName",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                {
                    "name": "x-token",
                    "in": "header",
                    "description": "Token needs to validate & authenticate",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "responses": {
                        "200": {
                            "description": "Success in getting list of employee for the searched string",
                            "schema": {
                                "$ref": "#/definitions/searchEmployeeListByNameSuccess"
                            }
                        },
                        "401": {
                            "description": "Failed to get the list of list of employee for the searched string",
                            "schema": {
                                "$ref": "#/definitions/searchEmployeeListByNameFailure"
                            }
                        }
                    }
                }
            }
        },
        "/searchEmployee/{searchKey}/{cursor}": {
            "get": {
                "tags": [
                    "Engineer", "Search"
                ],
                "description": "Search the engineer by search key & cursor, Search key matches the Alphabets 'A', 'B', 'C' or engineer name, city, company",
                "operationId": "searchEmployeeByKey",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                  {
                      "name": "searchKey",
                      "in": "path",
                      "description": "Search key will match to matches the Alphabets 'A', 'B', 'C' or engineer name, city & company",
                      "required": true,
                      "type": "string"
                  },
                  {
                      "name": "cursor",
                      "in": "path",
                      "description": "Cursor is the position for the redis. Initially it will be 0.",
                      "default":0,
                      "required": true,
                      "type": "string"
                  },
                  {
                      "name": "x-token",
                      "in": "header",
                      "description": "Token needs to validate & authenticate",
                      "required": true,
                      "type": "string"
                 }],
                "responses": {
                    "200": {
                        "description": "response",
                        "schema": {
                            "$ref": "#/definitions/searchEmployeeByKeySuccess"
                        }
                    },
                    "401": {
                        "description": "Failed to search employee for the search key",
                        "schema": {
                            "$ref": "#/definitions/searchEmployeeByKeyFailure"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "loginResponseSuccess": {
            "properties": {
                "status": {
                    "required": true,
                    "type": "Interger",
                    "example": true
                },
                "message": {
                    "required": true,
                    "type": "String",
                    "example": "User logged in login Successfully"
                },
                "token": {
                    "required": true,
                    "type": "String",
                    "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJhZG1pbkBicmlkZ2VsYWJ6LmNvbSIsImlhdCI6MTQ4Nzc2MDY3NSwiZXhwIjoxNDg4OTcwMjc1fQ.ALOhIMuqtczsrZkR-vtTbceP5djAQNk7CCwoZzKkOCY"
                }
            }
        },
        "loginResponseFailure": {
            "properties": {
                "status": {
                    "required": true,
                    "type": "Interger",
                    "example": false
                },
                "message": {
                    "required": true,
                    "type": "String",
                    "example": "Failed to authenticate user"
                }
            }
        },
        "DashboardResponseSuccess": {
            "properties": {
                "status": {
                    "required": true,
                    "type": "Interger",
                    "example": true
                },
                "message": {
                    "required": true,
                    "type": "String",
                    "example": "Successfully retrived the dashborad data"
                },
                "attendanceSummary": {
                    "type": "schema",
                    "example": {
                        "marked": 1,
                        "unmarked": "35"
                    }
                },
                "attendanceFallout": {
                    "type": "schema",
                    "example": {
                        "falloutEmployee": 35,
                        "totalEmployee": 36
                    }
                },
                "leaveSummary": {
                    "type": "schema",
                    "example": {
                        "leave": "4"
                    }
                }
            }
        },
        "DashboardResponseFailure": {
            "properties": {
                "status": {
                    "required": true,
                    "type": "Interger",
                    "example": false
                },
                "message": {
                    "required": true,
                    "type": "String",
                    "example": "Failed to retrive the dashborad data"
                }
            }
        },
...
        "attendanceSummaryByDateRangeSuccess": {
            "properties": {
                "status": {
                    "required": true,
                    "type": "Interger",
                    "example": true
                },
                "message": {
                    "required": true,
                    "type": "String",
                    "example": "Successfully retrived the attendance summary for the selected date range"
                },
                "attendance": {
                    "required": true,
                    "schema": {
                        "type": "array",
                        "example": [
                        {
                            "date": "02/03/2017",
                            "unmarked": "-"
                        },
                        {
                            "date": "02/03/2017",
                            "unmarked": "5"
                        }]
                    }
                },
                "totalEmployee": {
                    "required": true,
                    "type": "integer",
                    "example": 36
                }
            }
        },
        "attendanceSummaryByDateRangeFailure": {
            "properties": {
                "properties": {
                    "status": {
                        "required": true,
                        "type": "Interger",
                        "example": false
                    },
                    "message": {
                        "required": true,
                        "type": "String",
                        "example": "Failed to retrived the attendance summary for the selected date range"
                    }
                }
            }
        },
...
        "searchEmployeeListByNameFailure": {
            "properties": {
                "status": {
                    "required": true,
                    "type": "Interger",
                    "example": false
                },
                "message": {
                    "required": true,
                    "type": "String",
                    "example": "Failed to get the employee list for the searched employee name"
                }
            }
        },
        "recordEmployeeDayAttendanceSchema": {
            "properties": {
                "attendanceDate": {
                    "type": "string",
                    "required": true,
                    "example": "01/04/2017"
                },
                "engineerID": {
                    "type": "string",
                    "required": true,
                    "example": "40001EI"
                },
                "status": {
                    "type": "string",
                    "required": true,
                    "example": "Present"
                },
                "marked": {
                    "type": "boolean",
                    "required": true,
                    "example": true
                },
                "inTime": {
                    "type": "string",
                    "required": true,
                    "example": "8:30 AM"
                },
                "outTime": {
                    "type": "string",
                    "required": true,
                    "example": "7:30 PM"
                },
                "reason": {
                    "type": "string",
                    "required": true,
                    "example": "-"
                }
            }
        },
        "recordEmployeeDayAttendanceSuccess": {
            "properties": {
                "status": {
                    "required": true,
                    "type": "Interger",
                    "example": true
                },
                "message": {
                    "required": true,
                    "type": "String",
                    "example": "Successfully created employee day attendance record"
                },
                "engineerID": {
                    "required": true,
                    "type": "string",
                    "example": "40001EI"
                },
                "date": {
                    "required": true,
                    "type": "string",
                    "example": "01/04/2017"
                }
            }
        },
        "recordEmployeeDayAttendanceFailure": {
            "properties": {
                "status": {
                    "required": true,
                    "type": "Interger",
                    "example": false
                },
                "message": {
                    "required": true,
                    "type": "String",
                    "example": "Failed to create employee day attendance record"
                }
            }
        },
        "employeeMonthAttendanceSuccess": {
            "properties": {
                "status": {
                    "required": true,
                    "type": "Interger",
                    "example": true
                },
                "message": {
                    "required": true,
                    "type": "String",
                    "example": "Successfully retrived the employee month records"
                },
                "attendanceData": {
                    "required": true,
                    "type": "object",
                    "example": {
                        "1": {
                            "status": "Leave",
                            "marked": "true",
                            "intime": "-",
                            "outtime": "-",
                            "reason": "I don't care"
                        },
                        "29": {
                            "status": "Leave",
                            "marked": "true",
                            "intime": "-",
                            "outtime": "-",
                            "reason": "My Wish"
                        }
                    }
                },
                "employeeData": {
                    "required": true,
                    "type": "object",
                    "example": {
                        "name": "Amit R Singh",
                        "status": "Fellowship",
                        "company": "BridgeLabz",
                        "mobile": "9594368083",
                        "emailAddress": "xyz@gmail.com",
                        "BLStartDate": "1477308028",
                        "companyJoinDate": "1477308028",
                        "companyLeaveDate": "1477308028",
                        "leaveTaken": 7,
                        "imageUrl": "http://192.168.1.102:3000/image2.jpg"
                    }
                }
            }
        },
        "employeeMonthAttendanceFailure": {
            "properties": {
                "status": {
                    "required": true,
                    "type": "Interger",
                    "example": false
                },
                "message": {
                    "required": true,
                    "type": "String",
                    "example": "Failed to retrived the employee month records"
                }
            }
        },
...
        "generateAllEmployeeSalaryReportInvoiceSuccess": {
            "properties": {
                "status": {
                    "required": true,
                    "type": "Interger",
                    "example": true
                },
                "message": {
                    "required": true,
                    "type": "String",
                    "example": "Successfully retrived All Employee Salary Report & Invoice"
                },
                "allEmployee": {
                    "required": true,
                    "schema": {
                        "type": "array",
                        "example": [{
                                "engineerID": 30001,
                                "name": "Mrunal desai",
                                "status": "Internship",
                                "company": "BridgeLabz"
                            },
                            {
                                "engineerID": 30002,
                                "name": "Arati Patel",
                                "status": "Fellowship",
                                "company": "BridgeLabz"
                            }
                        ]
                    }
                }
            }
        },
        "generateAllEmployeeSalaryReportInvoiceFailure": {
            "properties": {
                "status": {
                    "required": true,
                    "type": "Interger",
                    "example": false
                },
                "message": {
                    "required": true,
                    "type": "String",
                    "example": "Failed to retrive All Employee Salary Report & Invoice"
                }
            }
        },
...
        "attendanceOfInternshipAtClientSuccess": {
            "properties": {
                "status": {
                    "required": true,
                    "type": "boolean",
                    "example": true
                },
                "message": {
                    "required": true,
                    "type": "string",
                    "example": "Successfully retrived the Attendance of Internship Employee at client side"
                },
                "allEmployee": {
                    "required": true,
                    "schema": {
                        "type": "array",
                        "example": [{
                            "company": "MERU",
                            "employeeList": {
                                "required": true,
                                "schema": {
                                    "type": "array",
                                    "example": [{
                                            "name": "AratiPatel",
                                            "leaves": 0
                                        },
                                        {
                                            "name": "Arjun Thakur",
                                            "leaves": 1
                                        }
                                    ]

                                }
                            }
                        },
                        {
                            "company": "NEXT-EDUCATION",
                            "employeeList": {
                                "required": true,
                                "schema": {
                                    "type": "array",
                                    "example": [{
                                            "name": "Sohail Khan",
                                            "leaves": 0
                                        },
                                        {
                                            "name": "Arbaz Khan",
                                            "leaves": 1
                                        }
                                    ]

                                }
                            }
                        }]
                    }
                }
            }
        },
        "attendanceOfInternshipAtClientFailure": {
            "properties": {
                "status": {
                    "required": true,
                    "type": "boolean",
                    "example": false
                },
                "message": {
                    "required": true,
                    "type": "string",
                    "example": "Failed to retrive the Attendance of Internship Employee at client side"
                }
            }
        },
...
        "searchEmployeeByKeyFailure": {
            "properties": {
                "status": {
                    "required": true,
                    "type": "Interger",
                    "example": false
                },
                "message": {
                    "required": true,
                    "type": "String",
                    "example": "Searched result failed"
                }
            }
        }
    }
}